refactor app structure
