+++ src/components/features/projects/DragPreview/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-const DragPreview = ({ name }) => {
+const Index = ({ name }) => {
   return (
     <div className="fixed pointer-events-none z-50 bg-background/95 backdrop-blur supports-[backdrop-filter]:bg-background/60 border rounded-md shadow-lg p-2 max-w-[200px]">
       <div className="flex items-center gap-2">
@@ -10,4 +10,4 @@ const DragPreview = ({ name }) => {
   );
 };
 
-export default DragPreview;
+export default Index;
